%
\name{Exponential.ANOVA.Quantiles}
\alias{Exponential.ANOVA.Quantiles}
\title{
Exponential quantile values
}
\description{
Quantiles for the exponential goodness-of-fit test. They are obtained 
from the sample correlation from ANOVA test for exponential distribution.
The number of Monte Carlo iterations is 1E08.

Dataset representing the quantiles and the associated critical values for the Weibullness
test. They were obtained by conducting Monte Carlo simulations where the
sample correlation coefficients were calculated based on ANOVA test for exponential distribution.
We used 1.0E08 Monte Carlo iterations in the simulation.
}
\usage{Exponential.ANOVA.Quantiles}
\format{
This data frame contains 998 rows and 1001 columns.
}
\keyword{datasets}
%% \keyword{exponential probability plot}% use one of  RShowDoc("KEYWORDS")
