%
\name{Gumbel.Plot.Quantiles}
\alias{Gumbel.Plot.Quantiles}
\title{
Gumbel quantile values
}
\description{
Quantiles for the Gumbel goodness-of-fit test. They are obtained 
from the sample correlation from the Gumbel probability plot. 
The number of Monte Carlo iterations is 1E08.


Dataset representing the quantiles and the associated critical values for the Gumbel 
goodness-of-fit test. They were obtained by conducting Monte Carlo simulations where the
sample correlation coefficients were calculated based on the Gumbel probability plot. 
We used 1.0E08 Monte Carlo iterations in the simulation.
}
\usage{Gumbel.Plot.Quantiles}
\format{
This data frame contains 998 rows and 1001 columns.
}
\keyword{datasets}
\keyword{Gumbel probability plot}% use one of  RShowDoc("KEYWORDS")
