%
\name{IW.Plot.Quantiles}
\alias{IW.Plot.Quantiles}
\alias{inverse.Weibull.Plot.Quantiles}
\title{
Inverse Weibull quantile values
}
\description{
Quantiles for the inverse Weibullness Test. They are obtained 
from the sample correlation from the inverse Weibull plot. 
The number of Monte Carlo iterations is 1E08.

Dataset representing the quantiles and the associated critical values for the inverse Weibullness
test. They were obtained by conducting Monte Carlo simulations where the
sample correlation coefficients were calculated based on the inverse Weibull plot. 
We used 1.0E08 Monte Carlo iterations in the simulation.
}
\usage{IW.Plot.Quantiles}
\format{
This data frame contains 998 rows and 1001 columns.
}
\references{
Park, C. (2017).
Weibullness test and parameter estimation of the three-parameter
Weibull model using the sample correlation coefficient. 
\emph{International Journal of Industrial Engineering - Theory, 
Applications and Practice} 
24(4), 376-391. \cr 
\doi{10.23055/ijietap.2017.24.4.2848}
}
\keyword{datasets}
\keyword{inverse Weibull plot}% use one of  RShowDoc("KEYWORDS")
