%
\name{Weibull.Plot.Quantiles}
\alias{Weibull.Plot.Quantiles}
\title{
Weibull quantile values
}
\description{
Quantiles for the Weibullness Test. They are obtained 
from the sample correlation from the Weibull plot. 
The number of Monte Carlo iterations is 1E08.


Data set representing the quantiles and the associated critical values for the Weibullness
test. They were obtained by conducting Monte Carlo simulations where the
sample correlation coefficients were calculated based on the Weibull plot. 
We used 1.0E08 Monte Carlo iterations in the simulation.
}
\usage{Weibull.Plot.Quantiles}
\format{
This data frame contains 998 rows and 1001 columns.
}
\references{
Park, C. (2017).
Weibullness test and parameter estimation of the three-parameter
Weibull model using the sample correlation coefficient. 
\emph{International Journal of Industrial Engineering - Theory, 
Applications and Practice} 
24(4), 376-391. \cr 
\url{http://journals.sfu.ca/ijietap/index.php/ijie/article/view/2848}
}
\keyword{datasets}
\keyword{Weibull plot}% use one of  RShowDoc("KEYWORDS")
